refactor: tag, attachment metadata helper
