+++ packages/backend/src/apps/gathersg/actions/get-case-details/get-data-out-metadata.ts
@@ -1,5 +1,9 @@
 import { IDataOutMetadata, IExecutionStep, IJSONArray } from '@plumber/types'
 
+import {
+  createAttachmentsMetadata,
+  createTagMetadata,
+} from '../../common/data-out-metadata-helpers'
 import { decodeFieldName } from '../../common/utils'
 
 import { dataOutSchema } from './schema'
@@ -61,29 +65,9 @@ async function getDataOutMetadata(
     },
   }
 
-  // handle tags if any
-  // tags are an array of strings and exist at the top level alongside the
-  // uuid, caseRef, etc
-  const tagsMetadata = Object.create(null)
-  if (dataOut.tags && Array.isArray(dataOut.tags)) {
-    for (let i = 0; i < dataOut.tags.length; i++) {
-      tagsMetadata[i] = { label: `Tag` }
-    }
-  }
-
-  const attachmentsMetadata = Object.create(null)
-  const attachmentKeys: string[] = []
-  if (dataOut.attachments) {
-    for (const key of Object.keys(dataOut.attachments)) {
-      attachmentsMetadata[key] = {
-        name: { isHidden: true },
-        mimeType: { isHidden: true },
-        size: { isHidden: true },
-      }
-
-      attachmentKeys.push(key)
-    }
-  }
+  const tagsMetadata = createTagMetadata(dataOut)
+  const { attachmentsMetadata, attachmentKeys } =
+    createAttachmentsMetadata(dataOut)
 
   // handle hex-encoded field names from dataOut
   const fieldsMetadata = Object.create(null)
